requirements.txt
Flask==2.0.2
Flask_SQLAlchemy==2.5.1
PyJWT==1.7.0
